Adding additional AES-GCM encryption as v2
This adds the AES-GCM encryption method which is needed for some devices on newer firmware versions (e.g. gree model 32776, v1.23).
The discovery happens automatically during the bind process with a retry on v2 if v1 fails.
Key and encryption handling was put into connection.js with little override possibilities to ensure automatic handling is working.
Additional logging was added.